2do com

+++ Persona.cs
@@ -2,5 +2,5 @@ public class Persona{
     int ID  { get; set;}
     string Nombre  { get; set;}
     string Direccion  { get; set;}
-    long int Telefono  { get; set;}
+     int Telefono  { get; set;}
 }
